+++ src/AuditTrail/Extenssions/RuleExtensions.cs
@@ -3,10 +3,10 @@ using AuditTrail.Fluent.Abstractions;
 using AuditTrail.Fluent.Rules;
 using System;
 
-namespace AuditTrail.Extenssions;
+namespace AuditTrail.Extensions;
 public static class RuleExtensions
 {
-    public static IRuleBulder<T, TPermission, TProperty?> Ignore<T, TPermission, TProperty>(this IRuleBulder<T, TPermission, TProperty> ruleBuilder)
+    public static IRuleBuilder<T, TPermission, TProperty?> Ignore<T, TPermission, TProperty>(this IRuleBuilder<T, TPermission, TProperty> ruleBuilder)
         where T : class
     {
         var rule = new IgnorePropertyRule<T, TProperty?>();
@@ -14,14 +14,14 @@ public static class RuleExtensions
     }
 
     /// <param name="includesHash"> Parameter important to pass exact same value that used to encrypt </param>
-    public static IRuleBulder<T, TPermission, TProperty?> Decrypt<T, TPermission, TProperty>(this IRuleBulder<T, TPermission, TProperty> ruleBuilder, IAuditTrailDecryption auditTrailDecryption, bool includesHash)
+    public static IRuleBuilder<T, TPermission, TProperty?> Decrypt<T, TPermission, TProperty>(this IRuleBuilder<T, TPermission, TProperty> ruleBuilder, IAuditTrailDecryption auditTrailDecryption, bool includesHash)
         where T : class
     {
         var rule = new DecryptPropertyRule<T, TProperty?>(auditTrailDecryption, includesHash);
         return ruleBuilder.SetRule(rule!)!;
     }
 
-    public static IRuleBulder<T, TPermission, TProperty?> ChangeName<T, TPermission, TProperty>(this IRuleBulder<T, TPermission, TProperty> ruleBuilder, string name)
+    public static IRuleBuilder<T, TPermission, TProperty?> ChangeName<T, TPermission, TProperty>(this IRuleBuilder<T, TPermission, TProperty> ruleBuilder, string name)
         where T : class
     {
         ArgumentNullException.ThrowIfNull(name);
@@ -29,7 +29,7 @@ public static class RuleExtensions
         return ruleBuilder.SetRule(rule!)!;
     }
 
-    public static IRuleBulder<T, TPermission, TProperty?> ChangeValue<T, TPermission, TProperty>(this IRuleBulder<T, TPermission, TProperty?> ruleBuilder, Func<TProperty?, object> func)
+    public static IRuleBuilder<T, TPermission, TProperty?> ChangeValue<T, TPermission, TProperty>(this IRuleBuilder<T, TPermission, TProperty?> ruleBuilder, Func<TProperty?, object> func)
         where T : class
     {
         ArgumentNullException.ThrowIfNull(func);