+++ src/AuditTrail/Fluent/Abstractions/IRuleBuilder.cs
@@ -1,0 +1,6 @@
+﻿namespace AuditTrail.Fluent.Abstractions;
+public interface IRuleBuilder<TEntity, TPermission, TProperty>
+    where TEntity : class
+{
+    IRuleBuilder<TEntity, TPermission, TProperty> SetRule(IPropertyRule<TEntity, TProperty> rule);
+}